Give the fan-out bound its own spec
The batching tests had landed in HoursApportionmentSpec, whose name
promises arithmetic. inBatches is a generic concurrency bound, not hours
math, so it gets a file that says so.
